refactor: split _operator.py into a package
Break the ~2400-line _operator.py into a _operator/ package:

- base.py: AbstractLinearOperator and the singledispatch operation stubs
- core.py: Matrix, PyTree, Jacobian, Function
- structured.py: Identity, Diagonal, Tridiagonal
- wrapper.py: Tagged, Tangent, Mul, Neg, Div
- binary.py: Add, Composed
- __init__.py re-exports the public API unchanged

Operation registrations are co-located with their operators (the stubs stay in
base.py), and repeated registrations are folded into loops (the core predicates
and the structured checks). Cross-file helpers drop their leading underscore,
and _try_sparse_materialise becomes try_structured_materialise. structured.py is
ordered Identity -> Diagonal -> Tridiagonal, and the operator API docs are
regrouped into core / structured / wrapper to match.

Behaviour-preserving: the public imports are unchanged, and the operator,
transpose, and adjoint test suites pass.
